general tidying.
- basic (very basic!) error handling
- moved to /api/
- handled update/delete invalid account cases
- added more account tests (not finished still)

+++ src/main/java/com/flatmate/flatmateregistry/AccountServiceImpl.java
@@ -3,12 +3,14 @@ package com.flatmate.flatmateregistry;
 import com.flatmate.flatmatepersistence.Account;
 import com.flatmate.flatmatepersistence.Transaction;
 import com.google.common.collect.Lists;
+import java.util.Objects;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
+
+import static com.flatmate.ErrorHandler.handleError;
 
 @Service
 public class AccountServiceImpl implements AccountService {
@@ -31,20 +33,31 @@ public class AccountServiceImpl implements AccountService {
     }
 
     public Account createAccount(final Account account) {
+        // TODO(alistair): validation?
         return accountRepository.save(account);
     }
 
     public Account updateAccount(final Long accountId, final Account account) {
-        account.setId(accountId);
-        return accountRepository.save(account);
+        if (Objects.nonNull(accountRepository.findOne(accountId))) {
+            account.setId(accountId);
+            return accountRepository.save(account);
+        } else {
+            // TODO(alistair): implement some sort of error handling here
+            handleError("Account not found!");
+            return null;
+        }
     }
 
     public void deleteAccount(final Long accountId) {
         final Account accountToDelete = accountRepository.findOne(accountId);
-        final Set<Transaction> transactions = accountToDelete.getTransactions();
-        transactions.forEach(transaction ->
-                transactionRepository.delete(transaction.getId()));
-        accountRepository.delete(accountId);
+        if (Objects.nonNull(accountToDelete)) {
+            final Set<Transaction> transactions = accountToDelete.getTransactions();
+            transactions.forEach(transaction ->
+                    transactionRepository.delete(transaction.getId()));
+            accountRepository.delete(accountId);
+        } else {
+            handleError("Account not found!");
+        }
     }
 
 }
